requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
tqdm>=4.62.0
pyyaml>=6.0
ujson>=5.0.0

# API clients (optional, install as needed)
openai>=1.12.0        # For OpenAI provider
ollama>=0.1.7         # For Ollama provider
httpx>=0.25.2         # For HTTP requests
aiohttp>=3.9.1        # For async HTTP requests

# HuggingFace dependencies (optional, install as needed)
torch>=2.0.0          # For HuggingFace provider
transformers>=4.30.0  # For HuggingFace provider
accelerate>=0.20.0    # For multi-GPU support
bitsandbytes>=0.40.0  # For quantization
sentencepiece>=0.1.99 # For some tokenizers
safetensors>=0.3.0    # For safe model loading

# Utilities
jsonschema>=4.0.0
pydantic>=2.0.0
scikit-learn>=1.0.0
python-Levenshtein==0.27.1